[FIX] base_tier_validation_forward: Tier review sequence should be an integer

+++ base_tier_validation_forward/models/tier_review.py
@@ -25,7 +25,7 @@ class TierReview(models.Model):
         compute="_compute_definition_data",
         store=True,
     )
-    sequence = fields.Float()
+    sequence = fields.Integer()
     has_comment = fields.Boolean(
         compute="_compute_definition_data",
         store=True,
